requirements.txt
feedparser==6.0.10
tweepy==4.15.0
facebook-sdk==3.1.0
linkedin-api==2.0.0a5
PyYAML==6.0.1
schedule==1.2.0
requests==2.31.0
beautifulsoup4==4.12.2
lxml==4.9.3 